Move JSONifying to a new thread
Previously JSONifying the LED data was done separately for each
websocket handler.  This took up 30% of a CPU core per websocket so is a
lot of load if multiple websockets are connected.  But we don't want to
do this in led_update which would burden the main thread with this.  So
make a new JSONifier thread which receives LedUpdates from the main
thread, turns them into JSON, then broadcasts them to all the websocket
handlers