Allow simulator to run on PCs
Add a rust "hardware" feature which you can disable to disable all of
the peripherals and raspberry-pi specifics.

When this feature is disabled all of the sensor values get stubbed out
and the LED updates are only sent to the web visualiser.  Also the
reporter is disabled to minimise the spam the backend sees.

Update the web visualiser to have button to decide whether it connects
to the actual hardware or to the simulator on localhost.